actualizarSistema ahora soporta Arch Linux

+++ actualizarSistema.sh
@@ -1,12 +1,21 @@
 #!/bin/bash
-sudo apt update
-sudo apt full-upgrade
-sudo apt autoremove
-sudo apt autoclean
 
-clear
+function imprimirMensaje {
+	echo "---- Sistema actualizado! ----"
+}
 
-echo "########################################"
-echo "###        SISTEMA ACTUALIZADO       ###"
-echo "########################################"
-
+if [ -f /etc/debian_version ]; then
+	sudo apt update
+	sudo apt full-upgrade
+	sudo apt autoremove
+	sudo apt autoclean
+	clear
+	imprimirMensaje
+elif [ -f /etc/arch-release ]; then
+	sudo pacman -Syu
+	clear
+	imprimirMensaje
+else
+	echo "Sistema no soportado"
+	exit 1
+fi
